# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
  '/api/users/{user}':
    get:
      summary: ''
      operationId: getApiUsersUser
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: user
        description: 'The user.'
        example: ab
        required: true
        schema:
          type: string
  /api/v1/login:
    post:
      summary: ''
      operationId: postApiV1Login
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: katarina67@example.net
                password:
                  type: string
                  description: 'Must be at least 6 characters.'
                  example: '.S/ZG#,efME0gXN/'
                full_name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: jopeiv
                photo_url:
                  type: string
                  description: ''
                  example: 'https://www.davis.biz/debitis-maiores-cupiditate-sequi-delectus-eos-incidunt'
              required:
                - email
                - password
      security: []
  /api/v1/register:
    post:
      summary: ''
      operationId: postApiV1Register
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: ekuhn@example.com
                password:
                  type: string
                  description: 'Must be at least 6 characters.'
                  example: d|T>QxzNqb>a%2b==v;
                full_name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: qvzvukyefowekwenfqph
                photo_url:
                  type: string
                  description: ''
                  example: 'http://white.com/quos-ut-laudantium-eos'
              required:
                - email
                - password
      security: []
...
  '/api/v1/users/{user}':
    get:
      summary: 'Display a specific user.'
      operationId: displayASpecificUser
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - 'User management'
      security: []
    delete:
      summary: 'Delete a specific user.'
      operationId: deleteASpecificUser
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  data:
                    message: 'User deleted successfully'
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'User deleted successfully'
      tags:
        - 'User management'
      security: []
    parameters:
      -
        in: path
        name: user
        description: 'The user.'
        example: eaque
        required: true
        schema:
          type: string
      -
        in: path
        name: id
        description: 'The ID of the user.'
        example: '1'
        required: true
        schema:
          type: string
...